requirements.txt
Django
django-crispy-forms
django-storages
azure-storage-blob
crispy-bootstrap4
sqlparse
tzdata
psycopg2-binary
whitenoise
Pillow
gunicorn